==> notebooks/3-cleaning-data.ipynb <==
import re
import pandas as pd
import matplotlib.pyplot as plt
from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory

## Import Data

DF_PATH = '../data/interim/'
df_food = pd.read_csv(DF_PATH + '0_food_labelled_resampled.csv')
df_electronic = pd.read_csv(DF_PATH + '0_electronic_labelled_resampled.csv')
df_fashion = pd.read_csv(DF_PATH + '0_fashion_labelled_resampled.csv')

df_food.head()

df_electronic.head()

df_fashion.head()

## Clean Reviews

sastrawi = StopWordRemoverFactory()
stopword = sastrawi.get_stop_words()
factory = StemmerFactory()
stemmer = factory.create_stemmer()

### Lower, Remove Newline, Stemming, & Remove Stopword

def clean_text_1(text):
    text = text.lower() # lower all sentence
    text = text.strip() # remove newline
    text = stemmer.stem(text) # Stemming kata berimbuhan menjadi kata dasar
    text = ' '.join(word for word in text.split() if word not in stopword)
    return text

print('Cleaning Food Reviews...')
df_food['clean_review_v1'] = df_food['reviews'].apply(clean_text_1)  
print('Cleaning Electronic Reviews...')
df_electronic['clean_review_v1'] = df_electronic['reviews'].apply(clean_text_1) 
print('Cleaning Fashion Reviews...') 
df_fashion['clean_review_v1'] = df_fashion['reviews'].apply(clean_text_1)
print('Cleaning Data Done.')  


df_food.to_csv('../data/interim/1_food_cleaned_v1.csv', index = False)
df_electronic.to_csv('../data/interim/1_electronic_cleaned_v1.csv', index = False)
df_fashion.to_csv('../data/interim/1_fashion_cleaned_v1.csv', index = False)

### Cleaning Character Repetition

df_food_v1 = pd.read_csv(DF_PATH + '1_food_cleaned_v1.csv')
df_electronic_v1 = pd.read_csv(DF_PATH + '1_electronic_cleaned_v1.csv')
df_fashion_v1 = pd.read_csv(DF_PATH + '1_fashion_cleaned_v1.csv')

def replace_two_or_more(s):
    #look for 2 or more repetitions of character
    pattern = re.compile(r"(.)\1{1,}", re.DOTALL) 
    return pattern.sub(r"\1", s)

def clean_text_2(text):
    text = ' '.join(replace_two_or_more(word) for word in str(text).split())
    return text

print('Cleaning Food Reviews...')
df_food_v1['clean_review_v2'] = df_food_v1['clean_review_v1'].apply(clean_text_2)  
print('Cleaning Electronic Reviews...')
df_electronic_v1['clean_review_v2'] = df_electronic_v1['clean_review_v1'].apply(clean_text_2) 
print('Cleaning Fashion Reviews...') 
df_fashion_v1['clean_review_v2'] = df_fashion_v1['clean_review_v1'].apply(clean_text_2)
print('Cleaning Data Done.')  

df_fashion_v1

### Menghapus Kata yang memiliki frekuensi <= 3

def count_word(df):
    counter = df.clean_review_v2.str.split(expand=True).stack().value_counts().reset_index()
    counter.columns = ['Word', 'Frequency'] 
    return counter

df_food_freq = count_word(df_food_v1)
food_freq_mino = list(df_food_freq[df_food_freq['Frequency'] <= 3]['Word'])

df_electronic_freq = count_word(df_electronic_v1)
electronic_freq_mino = list(df_electronic_freq[df_electronic_freq['Frequency'] <= 3]['Word'])

df_fashion_freq = count_word(df_fashion_v1)
fashion_freq_mino = list(df_fashion_freq[df_fashion_freq['Frequency'] <= 3]['Word'])

fashion_freq_mino

def clean_text_3(df, stopwords):
    list_text = []
    for text in df.clean_review_v2:
        list_text.append(' '.join(word for word in str(text).split() if word not in stopwords))
    df['clean_review_v3'] = list_text
    return df

print('Cleaning Food Reviews...')
df_food_v1 = clean_text_3(df_food_v1, food_freq_mino)
print('Cleaning Electronic Reviews...')
df_electronic_v1 = clean_text_3(df_electronic_v1, electronic_freq_mino)
print('Cleaning Fashion Reviews...')
df_fashion_v1 = clean_text_3(df_fashion_v1, fashion_freq_mino)
print('Cleaning Done.')

df_food_v1

### Menghapus kata dengan panjang 3 karakter
Kata dengan 3 karakter biasanya merupakan kata imbuhan ataupun kata sambung

def clean_text_4(text):
    text = ' '.join(word for word in str(text).split() if len(word) > 3)
    return text

print('Cleaning Food Reviews...')
df_food_v1['clean_review_v4'] = df_food_v1['clean_review_v3'].apply(clean_text_4)  
print('Cleaning Electronic Reviews...')
df_electronic_v1['clean_review_v4'] = df_electronic_v1['clean_review_v3'].apply(clean_text_4) 
print('Cleaning Fashion Reviews...') 
df_fashion_v1['clean_review_v4'] = df_fashion_v1['clean_review_v3'].apply(clean_text_4)
print('Cleaning Data Done.')  

df_food_v1

counter = df_fashion_v1.clean_review_v4.str.split(expand=True).stack().value_counts().reset_index()
counter.columns = ['Word', 'Frequency']
counter

df_food_v1.to_csv('../data/interim/2_food_cleaned_v2.csv', index = False)
df_electronic_v1.to_csv('../data/interim/2_electronic_cleaned_v2.csv', index = False)
df_fashion_v1.to_csv('../data/interim/2_fashion_cleaned_v2.csv', index = False)

### Menghapus Data Missing Value

df_food_v2 = pd.read_csv('../data/interim/2_food_cleaned_v2.csv')
df_electronic_v2 = pd.read_csv('../data/interim/2_electronic_cleaned_v2.csv')
df_fashion_v2 = pd.read_csv('../data/interim/2_fashion_cleaned_v2.csv')

df_food_v2 = df_food_v2[['clean_review_v4', 'label']]
df_electronic_v2 = df_electronic_v2[['clean_review_v4', 'label']]
df_fashion_v2 = df_fashion_v2[['clean_review_v4', 'label']]

#### Cek Missing Value

# Missing Value Data Food
df_food_v2.isna().sum()

# Missing Value Data Electronic
df_electronic_v2.isna().sum()

# Missing Value Data Fashion
df_fashion_v2.isna().sum()

#### Hapus Missing Value

df_food_v2 = df_food_v2.dropna(axis=0).reset_index(drop=True)
df_electronic_v2 = df_electronic_v2.dropna(axis=0).reset_index(drop=True)
df_fashion_v2 = df_fashion_v2.dropna(axis=0).reset_index(drop=True)

#### Cek Ulang Missing Value

# Missing Value Data Food
df_food_v2.isna().sum()

# Missing Value Data Electronic
df_electronic_v2.isna().sum()

# Missing Value Data Fashion
df_fashion_v2.isna().sum()

### Rename Kolom

df_food_v2 = df_food_v2.rename(columns = {'clean_review_v4' : 'reviews'})
df_electronic_v2 = df_electronic_v2.rename(columns = {'clean_review_v4' : 'reviews'})
df_fashion_v2 = df_fashion_v2.rename(columns = {'clean_review_v4' : 'reviews'})

df_food_v2

df_electronic_v2

df_fashion_v2

df_food_v2.to_csv('../data/interim/3_food_cleaned_v3.csv', index = False)
df_electronic_v2.to_csv('../data/interim/3_electronic_cleaned_v3.csv', index = False)
df_fashion_v2.to_csv('../data/interim/3_fashion_cleaned_v3.csv', index = False)
