==> notebooks/4-split-data.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from collections import Counter

DF_PATH = '../data/interim/'

## Import Data


df_food = pd.read_csv(DF_PATH + '3_food_cleaned_v3.csv')
df_electronic = pd.read_csv(DF_PATH + '3_electronic_cleaned_v3.csv')
df_fashion = pd.read_csv(DF_PATH + '3_fashion_cleaned_v3.csv')

df_food.head()

df_electronic.head()

df_fashion.head()

## Split Dataset

def split_data(df, test_size, target):
    X = df.drop(target, axis=1)  
    y = df[target]			

    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=42, stratify=y)
    df_train, df_test = pd.DataFrame(), pd.DataFrame()
    df_train['reviews'], df_test['reviews'] = X_train, X_test
    df_train['label'], df_test['label'] = y_train, y_test
    df_train, df_test = df_train.reset_index(drop=True), df_test.reset_index(drop=True)
    return df_train, df_test    

train_food, test_food = split_data(df_food, 0.2, 'label')
train_food, test_food

train_electronic, test_electronic = split_data(df_electronic, 0.2, 'label')
train_electronic, test_electronic

train_fashion, test_fashion = split_data(df_fashion, 0.2, 'label')
train_fashion, test_fashion

print('Rasio Label Data Training Kategori Food :', Counter(train_food.label))
print('Rasio Label Data Testing Kategori Food :',Counter(test_food.label))

## Save Data

train_food.to_csv('../data/processed/train/food_train.csv', index = False)
train_electronic.to_csv('../data/processed/train/electronic_train.csv', index = False)
train_fashion.to_csv('../data/processed/train/fashion_train.csv', index = False)

test_food.to_csv('../data/processed/test/food_test.csv', index = False)
test_electronic.to_csv('../data/processed/test/electronic_test.csv', index = False)
test_fashion.to_csv('../data/processed/test/fashion_test.csv', index = False)
