==> notebooks/5-training-model.ipynb <==
import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import TfidfVectorizer

## Import Data

def load_data(category):
    train = pd.read_csv('../data/processed/train/' + category + '_train.csv')
    test = pd.read_csv('../data/processed/test/' + category + '_test.csv')

    return train, test

food_train, food_test = load_data('food')
electronic_train, electronic_test = load_data('electronic')
fashion_train, fashion_test = load_data('fashion')


food_train.isna().sum()

food_train = food_train.dropna(axis=0)

## Train Model

### Train Food Model

X_train, y_train = food_train['reviews'], food_train['label']
X_test, y_test = food_test['reviews'], food_test['label']


vectorizer = TfidfVectorizer(ngram_range=(1,2), min_df=10)
features = vectorizer.fit_transform(X_train)

def trainingData(X, y):
    #membangun vector space model/pembobotan dengan tfidf
    vectorizer = TfidfVectorizer(ngram_range=(1,2), min_df=10)
    features = vectorizer.fit_transform(X)

    #melakukan split data training untuk mengetahui akurasi
    X_train, X_test, y_train, y_test = train_test_split(features, y, test_size=0.1, random_state=4)

    #modeling sentiment
    LR_ = LogisticRegression(C=3, solver='liblinear', max_iter=150).fit(X_train, y_train)

    # melakukan evaluasi
    yhat = LR_.predict(X_test)

    yhat_prob = LR_.predict_proba(X_test)
    
    return LR_, vectorizer
